requirements.txt
mcp[cli]
requests
google-api-python-client
google-auth-oauthlib
openai
python-dotenv
pydantic